requirements.txt
altgraph==0.17.4
packaging==25.0
pillow==11.2.1
PyGObject==3.42.1
pyinstaller-hooks-contrib==2025.5
pyinstaller==6.14.1
pystray==0.19.5
python-xlib==0.29
six==1.16.0